Fix installation routine in Makefile
There is an issue while performing the installation from scratch:

Postfix's ``master.cf`` and ``main.cf`` are generated during the
installation process of Postfix. So, their timestamp corresponds to the
time of running the installation.

It must be ensured, that **mailsrv**'s configuration files have a more
recent timestamp, otherwise ``make`` will not trigger the required
recipe to copy **mailsrv**'s configs to ``/etc/postfix``.

This *may* be easily fixed by ``touching`` the generated files, but
preferably this should work out-of-the-box.